refactor(issue triage): isolate issue meta and issue triage.
